new db, some fix
more warehouse

+++ AI-Integrated Construction Project Management System/cpms_Application/Interfaces/IAuthService.cs
@@ -12,7 +12,7 @@ namespace cpms_Application.Interfaces
     {
         Task<ApiResponse> RegisterAsync(UserRegisterRequest userRequest);
         Task<ApiResponse> LoginAsync(LoginRequest request);
-        Task<ApiResponse> VerifyEmailAsync(int userId, string verificationCode);
+        Task<ApiResponse> VerifyEmailAsync(long userId, string verificationCode);
 
     }
 }
